AIMS-252: -Moved the remaning_matches method from Request to a RequestQuery object
